server/requirements.txt
fastapi==0.115.6
pydantic==2.10.3
openai==1.57.0
aiohttp==3.11.10
beautifulsoup4==4.12.3
python-dotenv==1.0.1
uvicorn==0.32.1
pyodbc==5.2.0
azure-identity==1.19.0
psycopg2-binary==2.9.10